+++ src/features/advanced/index.ts
@@ -6,7 +6,7 @@ import './bs-hide-zero-workforce';
 import './cogcpex-clean-labels';
 import './context-controls-no-hover';
 import './cxos-hide-exchange';
-import './cxpo-auto-pricing';
+import './cxpo-auto-price';
 import './cxpo-shorten-fields';
 import './finla-hide-ecd';
 import './flt-flight-status-icons';
